+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -78,6 +78,18 @@ exclude: true
 
 [68.16]: ../GeckoRuntimeSettings.Builder.html#configFilePath-java.lang.String-
 
+- Added [`unregisterWebExtension`][68.17] to unregister a web extension.
+
+[68.17]: ../GeckoRuntime.html#unregisterWebExtension-org.mozilla.geckoview.WebExtension-
+
+- Added messaging support for WebExtension. [`setMessageDelegate`][68.18]
+  allows embedders to listen to messages coming from a WebExtension.
+  [`Port`][68.19] allows bidirectional communication between the embedder and
+  the WebExtension.
+
+[68.18]: ./WebExtension.html#setMessageDelegate-org.mozilla.geckoview.WebExtension.MessageDelegate-java.lang.String-
+[68.19]: ./WebExtension.Port.html
+
 ## v67
 - Added [`setAutomaticFontSizeAdjustment`][67.2] to
   [`GeckoRuntimeSettings`][67.3] for automatically adjusting font size settings
@@ -284,4 +296,4 @@ exclude: true
 [65.24]: ../CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: ../GeckoResult.html
 
-[api-version]: 77512ba4107f1bd3330327d8b9c1b91363430783
+[api-version]: 190a68155b36619b84113d822cebdf47e58f3599